mockstore: move mockstore/cluster to tikv/mockstore/cluster
