+++ store/tikv/delete_range_test.go
@@ -20,8 +20,8 @@ import (
 	"sort"
 
 	. "github.com/pingcap/check"
-	"github.com/pingcap/tidb/store/mockstore/cluster"
 	"github.com/pingcap/tidb/store/mockstore/mocktikv"
+	"github.com/pingcap/tidb/store/tikv/mockstore/cluster"
 )
 
 type testDeleteRangeSuite struct {
